Line for each opcode is stored more efficiently
We now use an array of (line, opcode offset) pairs which only tracks the
highest opcode offset for a given line.

This saves space when there are a large number of sequential opcodes
which are associated with the same line. The worst case stores as much
information as the previous implementation, if there are no sequential
opcodes that are associated to the same line.

Lookup time to determine the line number associated with a given opcode
offset has changed from constant to linear in the size of the array
of (line, opcode offset) pairs.